Final Backend Version

+++ Comments.Api/Models/DTOs/ApplicationUserDto.cs
@@ -9,6 +9,6 @@ namespace Comments.Api.Models.DTOs
         public required string Email { get; set; }
         public required string UserName { get; set; }
         public required string PhoneNumber { get; set; }
-        public int Score { get; set; } = 0;
+        public float Score { get; set; } = 0;
     }
 }
