+++ Comments.Api/Models/Domain/ApplicationUser.cs
@@ -6,6 +6,6 @@ namespace Comments.Api.Models.Domain
     {
         public required string FirstName { get; set; }
         public required string LastName { get; set; }
-        public int Score { get; set; } = 0;
+        public float Score { get; set; } = 0;
     }
 }
