+++ Gateway.Api/Models/ApplicationUser.cs
@@ -6,7 +6,7 @@ namespace Gateway.Api.Models
     {
         public required string FirstName { get; set; }
         public required string LastName { get; set; }
-        public int Score { get; set; } = 0;
+        public float Score { get; set; } = 0;
         public Ban? Ban { get; set; }
         public ICollection<Comment> Comments { get; set; } = new List<Comment>();
         public ICollection<Vote> Votes { get; set; } = new List<Vote>();
