+++ Votes.Api/Models/DTOs/ApplicationUserDto.cs
@@ -7,6 +7,6 @@ namespace Votes.Api.Models.Domain
         public required string FirstName { get; set; }
         public required string LastName { get; set; }
         public required string UserName { get; set; }
-        public int Score { get; set; } = 0;
+        public float Score { get; set; } = 0;
     }
 }
